+++ FDI.Simple/AppIG4/ProductAppIG4Item.cs
@@ -136,7 +136,7 @@ namespace FDI.Simple
         public int ID { get; set; }
         public string Name { get; set; }
         
-        public decimal? Value { get; set; }
+        public double? Value { get; set; }
         public int? Quantity { get; set; }
     }
     public class ModelProductAppIG4Item : BaseModelSimple
